+++ lib/main.dart
@@ -24,7 +24,7 @@ import 'package:kino_bar_app/pages/wechselgeld_pruefen_seite.dart';
 import 'package:kino_bar_app/pages/datenschutz_seite.dart';
 import 'package:kino_bar_app/pages/kurzeinstieg_seite.dart';
 import 'package:kino_bar_app/pages/ueber_entwickler_seite.dart';
-import 'package:kino_bar_app/services/sw_update_service.dart';
+import 'package:kino_bar_app/services/update_lifecycle_watcher.dart';
 import 'package:kino_bar_app/services/update_reload_guard.dart';
 import 'package:kino_bar_app/storage/lokaler_speicher.dart';
 import 'package:kino_bar_app/theme/app_farben.dart';
@@ -70,13 +70,7 @@ Future<void> main() async {
   await WechselgeldConfigService().initOnAppStart();
 
   runApp(const MeineApp());
-  initSwUpdateWatcher(() {
-    if (!UpdateReloadGuard.istAufSichererSeite) {
-      return false;
-    }
-    reloadPage();
-    return true;
-  });
+  UpdateLifecycleWatcher().starten();
 }
 
 class MeineApp extends StatelessWidget {